Animate conveyor belt lines
Make the conveyor belt lines move across the belt.

Instead of moving each line along the length of belt and dealing with
resetting their positions after they  reach the end, each line is only
moved 30 pixels. Once it has moved 30 pixels it goes back to its
initial position.

This gives the illusion of the lines moving along the length of the
belt as by the time one belt resets its position, the previous belt
will have been in that position during the previous frame so the
transition is seamless.

Use a stencil to prevent the lines from drawing when they're no longer
on the belt.