Simplify conveyorOffset equation
Add local conveyorMovement variable to love.update() to set the speed
of the conveyor. It is added to the current conveyorOffset to
determine the next offset.

This change will allow simpler equations to be used in the future for
movement that depends on the conveyor, instead of using the math over
and over again.