Change presentOffset
Change presentOffset to decrement starting from 0

The math for calculating the positions of the presents was complicated
because their initial positions were set relative to ( 0, 0 ), in the
middle of the conveyor belt. This required a strange range for
presentOffset.

Change the initial positions to be relative to the right side of the
screen. This allows presentOffset to start at 0 and be added to the
initial positions.